+++ src/exec/GlbExecutor.ts
@@ -7,11 +7,11 @@ export class GlbExecutor extends SimpleExecutor {
         return type === 'glb';
     }
 
-    async execute(module: MiniModule): Promise<void> {
-        for (const file of module.tmpFiles!) {
+    async execute(context: BatchContext): Promise<void> {
+        for (const file of context.files) {
             // 将配置挂载到 pluto.assets[模块名]
             const blobUrl = base64ToBlobUrl(file.content, `model/gltf-binary`);
-            pluto.third.assets[module.name][file.name] = blobUrl;
+            pluto.third.assets[context.name][file.name] = blobUrl;
         };
     }