+++ src/exec/PageExecutor.ts
@@ -6,10 +6,10 @@ export class PageExecutor extends SimpleExecutor {
         return type === 'page';
     }
         
-    async execute(module: MiniModule): Promise<void> {
-        for (const file of module.tmpFiles!) {
+    async execute(context: BatchContext): Promise<void> {
+        for (const file of context.files) {
             // 将配置挂载到 pluto.assets[模块名]
-            pluto.third.assets[module.name][file.name] = file.content;
+            pluto.third.assets[context.name][file.name] = file.content;
         };
     }
 }