+++ src/manager/FormManager.ts
@@ -1,4 +1,4 @@
-import { FormModalConfig, FormValues } from "types/form";
+import { FormModalConfig, FormSubmitResult, FormValues } from "types/form";
 import { PlutoFormModal } from "modal/PlutoFormModal";
 import { FormJson } from "../modal/FormJson";
 
@@ -6,38 +6,26 @@ export class FormManager {
 
     async prompt(name: string, required: boolean = true, title: string = ''): Promise<string> {
         const config = FormJson.input(title, name, required);
-        return this.openJson(config, {})
-            .then((values) => values["name"] as string);
+        return this.openForm(config, {})
+            .then((result) => result.data?.["name"] as string);
     }
 
     async create(module: any): Promise<string> {
         const config = FormJson.create();
-        return this.openJson(config, module)
-            .then((values) => values["name"] as string);
+        return this.openForm(config, module)
+            .then((result) => result.data?.["name"] as string);
     }
 
-    async openSetting(module: any): Promise<FormValues> {
+    async openSetting(module: any): Promise<FormSubmitResult> {
         const config = FormJson.setting(module.name);
-        return this.openJson(config, module as FormValues)
-            .then((values) => values);
+        return this.openForm(config, module as FormValues);
     }
 
-    async openYaml(config: FormModalConfig, defaultValues: FormValues = {}): Promise<FormValues> {
+    async openForm(config: FormModalConfig, defaultValues: FormValues = {}): Promise<FormSubmitResult> {
         return new Promise((resolve, reject) => {
             new PlutoFormModal(config, defaultValues, (values) => {
                 console.log('表单提交值:', values);
-                resolve(values);
-            }, () => {
-                reject({ type: 'cancel', message: '用户取消了操作' });
-            }).open();
-        });
-    }
-
-    async openJson(config: FormModalConfig, defaultValues: FormValues = {}): Promise<FormValues> {
-        return new Promise((resolve, reject) => {
-            new PlutoFormModal(config, defaultValues, (values) => {
-                console.log('表单提交值:', values);
-                resolve(values);
+                resolve({status: 'ok', data: values});
             }, () => {
                 reject({ type: 'cancel', message: '用户取消了操作' });
             }).open();